ArmPkg/ArmMmuLib: Explicitly cast UINT32 data conversions
(cherry picked from commit 2bdc90429a149af6c0e88506730123586b6e08ae)

REF:https://bugzilla.tianocore.org/show_bug.cgi?id=2835

There's several occurrences of a UINT64 or an EFI_PHYSICAL_ADDRESS
being assigned to a UINT32 value in ArmMmuLib. These result in
warning C4244 in VS2019:

warning C4244: '=': conversion from 'UINT64' to 'UINT32', possible
loss of data

warning C4244: '=': conversion from 'EFI_PHYSICAL_ADDRESS' to
'UINT32', possible loss of data

This change explicitly casts the values to UINT32.

These can be reproduced with the following build command:
build -b DEBUG -a ARM -t VS2019 -p ArmPkg/ArmPkg.dsc
-m ArmPkg/Library/ArmMmuLib/ArmMmuBaseLib.inf

(cherry picked from commit 2bdc90429a149af6c0e88506730123586b6e08ae)

+++ ArmPkg/Library/ArmMmuLib/Arm/ArmMmuLibUpdate.c
@@ -152,7 +152,7 @@ UpdatePageEntries (
   FirstLevelTable = (ARM_FIRST_LEVEL_DESCRIPTOR *)ArmGetTTBR0BaseAddress ();
 
   // Calculate number of 4KB page table entries to change
-  NumPageEntries = Length / TT_DESCRIPTOR_PAGE_SIZE;
+  NumPageEntries = (UINT32)(Length / TT_DESCRIPTOR_PAGE_SIZE);
 
   // Iterate for the number of 4KB pages to change
   Offset = 0;
@@ -288,7 +288,7 @@ UpdateSectionEntries (
   ASSERT (FirstLevelIdx < TRANSLATION_TABLE_SECTION_COUNT);
 
   // calculate number of 1MB first level entries this applies to
-  NumSections = Length / TT_DESCRIPTOR_SECTION_SIZE;
+  NumSections = (UINT32)(Length / TT_DESCRIPTOR_SECTION_SIZE);
 
   // iterate through each descriptor
   for(i=0; i<NumSections; i++) {
